+++ Vungle/VungleAdapterConfiguration.m
@@ -44,7 +44,7 @@ typedef NS_ENUM(NSInteger, VungleAdapterErrorCode) {
 #pragma mark - MPAdapterConfiguration
 
 - (NSString *)adapterVersion {
-    return @"6.9.0.0";
+    return @"6.9.1.0";
 }
 
 - (NSString *)biddingToken {