+++ Vungle/CHANGELOG.md
@@ -20,6 +20,7 @@
 
 * 6.8.0.0
     * This version of the adapters has been certified with Vungle 6.8.0 and MoPub SDK 5.13.1.
+    * Remove VungleSDKResetPlacementForDifferentAdSize error check for loading Ads.
 
 * 6.7.1.0
     * This version of the adapters has been certified with Vungle 6.7.1 and MoPub SDK 5.13.1.