+++ Vungle/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Changelog
 * 6.9.0.0
     * This version of the adapters has been certified with Vungle 6.9.0 and MoPub SDK 5.15.0.
+    * Remove VungleSDKResetPlacementForDifferentAdSize error check for loading Ads.
 
 * 6.8.1.3
     * Support `orientations` data passing through MoPub initialize in iOS and Unity iOS.
@@ -20,7 +21,6 @@
 
 * 6.8.0.0
     * This version of the adapters has been certified with Vungle 6.8.0 and MoPub SDK 5.13.1.
-    * Remove VungleSDKResetPlacementForDifferentAdSize error check for loading Ads.
 
 * 6.7.1.0
     * This version of the adapters has been certified with Vungle 6.7.1 and MoPub SDK 5.13.1.