+++ Vungle/CHANGELOG.md
@@ -2,6 +2,7 @@
 * 6.9.0.0
     * This version of the adapters has been certified with Vungle 6.9.0 and MoPub SDK 5.15.0.
     * Remove VungleSDKResetPlacementForDifferentAdSize error check for loading Ads.
+    * Fix the issue that fullscreen ads fail to load.
 
 * 6.8.1.3
     * Support `orientations` data passing through MoPub initialize in iOS and Unity iOS.